Move Marvin media types to their own class
