+++ src/main/java/com/pinkpony/config/MarvinMediaTypes.java
@@ -1,0 +1,8 @@
+package com.pinkpony.config;
+
+import org.springframework.http.MediaType;
+
+public class MarvinMediaTypes {
+    public static final MediaType MARVIN_JSON_MEDIATYPE = new MediaType("application", "x-marvin-bot+json");
+    public static final String MARVIN_JSON_MEDIATYPE_VALUE = "application/x-marvin-bot+json";
+}
